requirements.txt
python-dotenv==0.10.3
requests==2.22.0
python-telegram-bot==11.1.0
redis==3.2.1
